feat: Add SwipableView component
This component allows swiping between exercise sets on the dashboard.  It uses `framer-motion` for animations and `react-swipeable` for swipe gesture detection.  The swipe threshold dynamically adjusts based on viewport width.
